Removed unnecessary directories and moved all jsx files with components directly to the components directory.
