+++ src/App.jsx
@@ -1,6 +1,6 @@
 import "./App.css";
-import Header from "./components/Header/Header.jsx";
-import Plans from "./components/Plans/Plans.jsx";
+import Header from "./components/Header.jsx";
+import Plans from "./components/Plans.jsx";
 import React from "react";
 
 function App() {
