+++ src/components/PartsOfPlan.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import PartsOfPlanInput from "../PartsOfPlanInput/PartsOfPlanInput.jsx";
-import {useInputLogic} from "../../hooks/useInputLogic.js";
-import {PARTS_OF_PLAN_KEYS} from "../../constants/keys.js";
-import PartsOfPlanData from "../PartsOfPlanData/PartsOfPlanData.jsx";
+import PartsOfPlanInput from "./PartsOfPlanInput.jsx";
+import {useInputLogic} from "../hooks/useInputLogic.js";
+import {PARTS_OF_PLAN_KEYS} from "../constants/keys.js";
+import PartsOfPlanData from "./PartsOfPlanData.jsx";
 
 const PartsOfPlan = () => {
     const {
